phase-3/backend/pyproject.toml
[project]
name = "todo-chatbot-backend"
version = "3.0.0"
description = "Phase III AI Chatbot backend - conversational todo management with MCP + OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Todo App Team"}
]

dependencies = [
    "fastapi>=0.128.0",
    "sqlmodel>=0.0.31",
    "alembic>=1.13.0",
    "asyncpg>=0.31.0",
    "pyjwt[crypto]>=2.11.0",
    "PyNaCl>=1.5.0",
    "cryptography>=46.0.4",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.9",
    "openai-agents>=0.7.0",
    "mcp>=1.26.0",
    "openai>=1.60.0",
    "apscheduler>=3.11.2",
    "pywebpush>=2.2.0",
    "py-vapid>=1.9.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.28.1",
    "ruff>=0.14.14",
    "mypy>=1.19.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["S101"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = "agents.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "mcp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "apscheduler.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pywebpush.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "jwt.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlmodel.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "phase2_backend.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "phase3_backend.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "py_vapid.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "httpx.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]
asyncio_mode = "auto"
